Move LinkedEditingRangeHelper functionality to its callers
